requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
python-multipart
torch
transformers
bitsandbytes
accelerate
python-dotenv
